backend/requirements.txt
pydantic==2.5.2 
fastapi
uvicorn[standard]
python-dotenv
python-docx
requests
cryptography
requests
textstat==0.7.3
python-docx